Extracted CollectionTask.MarkNoteMulti to markNoteMulti
